backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
asyncpg==0.30.0
alembic==1.16.2
psycopg2-binary==2.9.9
# If you are on macOS with Python 3.13 and the above fails, comment it and uncomment the line below
# psycopg

# Pydantic
pydantic[email]

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Configuration
python-dotenv==1.0.0

# For utilities
loguru

# Additional utilities
httpx<0.25.0,>=0.24.0

# Rate limiting
slowapi==0.1.9

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Logging and monitoring
prometheus-client==0.19.0

# Additional FastAPI extensions
fastapi-limiter==0.1.6
fastapi-cache2==0.2.1

# User management dependencies
bcrypt==3.2.0
PyJWT==2.8.0
email-validator==2.1.0
pydantic-settings==2.1.0
python-multipart==0.0.6
greenlet==3.2.3 